refactor: update PR comment keybinding from 'cc' to 'ca'
I have noticed that default proposed way to add normal comment is
`<leader>ua` (`a` stands for add and is idiomatic vim key to add
something), but previously `ca` was used to approved and `cc` to add PR
level comment. This was unintuitive and I accidentally approved PR
instead of commenting on it. Therefore approval is changed to `cA` and
this freed `ca` for adding comment. As well later it will be possible to
add `cd` to delete PR level comments and `cc` to change existing PR level
comments.

Note: this is defaults and you can change your key bindings in config if
default is not working for you.